Move NooPBilling to it's own class
